Add a style.css file linked to base.html
I added a style.css file to base.html, and also linked font from google font in it.
Edited play.html to change the h3 tag on the link, as it made more sense to have a p

The style.css seem to be misinterpreted by the browser.. Don't know why.