Added likes functionality with number of likes stored in a db along with videos' filenames.
This should be useful for implementing the comments because these can be stored also in the db and just like the two models (video and comment).
Like/dislike buttons are sending an asynchronous request (via jquery/ajax) in flask/media_server.py with a corresponding string indicating like or dislike. Flask receives them and updates the database num of likes (row) for the current video (column).
The video object stored in the database is stored in models.py. Other objects to be stored in the db could be added there too (don't forget to apply db migrations when making changes in the models).

Because the both media_server.py and models.py need a reference to the db -which is initialized within media_server.py- we create it in globalimports.py which we later import everwhere else to avoid circularity of SQLAlchemy's imports.

New dependencies where needed (flask-sqlalchemy, flask-migrate) which are listed in requirements.py.

Further improvments:
-limit even more the requests to the db by not commiting on every like incement/decrement
-add an likes counter inside the video page
-style buttons